move postgres class
